Optimize performance of pattern match cache
A profile of the JATS example showed a 30% overall time in the `evaluateXPathToString` function,
which was almost all in generate-id. Which is used to construct a cache key for nodes and
patterns. Replacing the use of generate-id with a nested map for the cache makes it disappear from
the profile!